requirements.txt
streamlit
numpy
tensorflow
opencv-python
streamlit-webrtcbrtc